chore: move assertion parsing into submodule
