+++ compiler/noirc_frontend/src/parser/parser.rs
@@ -37,17 +37,19 @@ use crate::lexer::Lexer;
 use crate::parser::{force, ignore_then_commit, statement_recovery};
 use crate::token::{Attribute, Attributes, Keyword, SecondaryAttribute, Token, TokenKind};
 use crate::{
-    BinaryOp, BinaryOpKind, BlockExpression, ConstrainKind, ConstrainStatement, Distinctness,
-    ForLoopStatement, ForRange, FunctionDefinition, FunctionReturnType, FunctionVisibility, Ident,
-    IfExpression, InfixExpression, LValue, Lambda, Literal, NoirFunction, NoirStruct, NoirTrait,
-    NoirTraitImpl, NoirTypeAlias, Param, Path, PathKind, Pattern, Recoverable, Statement,
-    TraitBound, TraitImplItem, TraitItem, TypeImpl, UnaryOp, UnresolvedTraitConstraint,
+    BinaryOp, BinaryOpKind, BlockExpression, Distinctness, ForLoopStatement, ForRange,
+    FunctionDefinition, FunctionReturnType, FunctionVisibility, Ident, IfExpression,
+    InfixExpression, LValue, Lambda, Literal, NoirFunction, NoirStruct, NoirTrait, NoirTraitImpl,
+    NoirTypeAlias, Param, Path, PathKind, Pattern, Recoverable, Statement, TraitBound,
+    TraitImplItem, TraitItem, TypeImpl, UnaryOp, UnresolvedTraitConstraint,
     UnresolvedTypeExpression, UseTree, UseTreeKind, Visibility,
 };
 
 use chumsky::prelude::*;
 use iter_extended::vecmap;
 use noirc_errors::{Span, Spanned};
+
+mod assertion;
 
 /// Entry function for the parser - also handles lexing internally.
 ///
@@ -792,9 +794,9 @@ where
 {
     recursive(|statement| {
         choice((
-            constrain(expr_parser.clone()),
-            assertion(expr_parser.clone()),
-            assertion_eq(expr_parser.clone()),
+            assertion::constrain(expr_parser.clone()),
+            assertion::assertion(expr_parser.clone()),
+            assertion::assertion_eq(expr_parser.clone()),
             declaration(expr_parser.clone()),
             assignment(expr_parser.clone()),
             for_loop(expr_no_constructors, statement),
@@ -806,64 +808,6 @@ where
 
 fn fresh_statement() -> impl NoirParser<StatementKind> {
     statement(expression(), expression_no_constructors(expression()))
-}
-
-fn constrain<'a, P>(expr_parser: P) -> impl NoirParser<StatementKind> + 'a
-where
-    P: ExprParser + 'a,
-{
-    ignore_then_commit(
-        keyword(Keyword::Constrain).labelled(ParsingRuleLabel::Statement),
-        expr_parser,
-    )
-    .map(|expr| StatementKind::Constrain(ConstrainStatement(expr, None, ConstrainKind::Constrain)))
-    .validate(|expr, span, emit| {
-        emit(ParserError::with_reason(ParserErrorReason::ConstrainDeprecated, span));
-        expr
-    })
-}
-
-fn assertion<'a, P>(expr_parser: P) -> impl NoirParser<StatementKind> + 'a
-where
-    P: ExprParser + 'a,
-{
-    let argument_parser =
-        expr_parser.separated_by(just(Token::Comma)).allow_trailing().at_least(1).at_most(2);
-
-    ignore_then_commit(keyword(Keyword::Assert), parenthesized(argument_parser))
-        .labelled(ParsingRuleLabel::Statement)
-        .validate(|expressions, span, _| {
-            let condition = expressions.first().unwrap_or(&Expression::error(span)).clone();
-            let message = expressions.get(1).cloned();
-            StatementKind::Constrain(ConstrainStatement(condition, message, ConstrainKind::Assert))
-        })
-}
-
-fn assertion_eq<'a, P>(expr_parser: P) -> impl NoirParser<StatementKind> + 'a
-where
-    P: ExprParser + 'a,
-{
-    let argument_parser =
-        expr_parser.separated_by(just(Token::Comma)).allow_trailing().at_least(2).at_most(3);
-
-    ignore_then_commit(keyword(Keyword::AssertEq), parenthesized(argument_parser))
-        .labelled(ParsingRuleLabel::Statement)
-        .validate(|exprs: Vec<Expression>, span, _| {
-            let predicate = Expression::new(
-                ExpressionKind::Infix(Box::new(InfixExpression {
-                    lhs: exprs.first().unwrap_or(&Expression::error(span)).clone(),
-                    rhs: exprs.get(1).unwrap_or(&Expression::error(span)).clone(),
-                    operator: Spanned::from(span, BinaryOpKind::Equal),
-                })),
-                span,
-            );
-            let message = exprs.get(2).cloned();
-            StatementKind::Constrain(ConstrainStatement(
-                predicate,
-                message,
-                ConstrainKind::AssertEq,
-            ))
-        })
 }
 
 fn declaration<'a, P>(expr_parser: P) -> impl NoirParser<StatementKind> + 'a
@@ -1671,13 +1615,19 @@ fn literal() -> impl NoirParser<ExpressionKind> {
 }
 
 #[cfg(test)]
-mod test {
+mod parser_test_helpers {
+    use chumsky::primitive::just;
+    use chumsky::Parser;
+    use iter_extended::vecmap;
     use noirc_errors::CustomDiagnostic;
 
-    use super::*;
-    use crate::{ArrayLiteral, Literal};
+    use crate::{
+        lexer::Lexer,
+        parser::{force, NoirParser},
+        token::Token,
+    };
 
-    fn parse_with<P, T>(parser: P, program: &str) -> Result<T, Vec<CustomDiagnostic>>
+    pub(crate) fn parse_with<P, T>(parser: P, program: &str) -> Result<T, Vec<CustomDiagnostic>>
     where
         P: NoirParser<T>,
     {
@@ -1691,7 +1641,10 @@ mod test {
             .map_err(|errors| vecmap(errors, Into::into))
     }
 
-    fn parse_recover<P, T>(parser: P, program: &str) -> (Option<T>, Vec<CustomDiagnostic>)
+    pub(crate) fn parse_recover<P, T>(
+        parser: P,
+        program: &str,
+    ) -> (Option<T>, Vec<CustomDiagnostic>)
     where
         P: NoirParser<T>,
     {
@@ -1704,7 +1657,7 @@ mod test {
         (opt, errors)
     }
 
-    fn parse_all<P, T>(parser: P, programs: Vec<&str>) -> Vec<T>
+    pub(crate) fn parse_all<P, T>(parser: P, programs: Vec<&str>) -> Vec<T>
     where
         P: NoirParser<T>,
     {
@@ -1720,7 +1673,7 @@ mod test {
         })
     }
 
-    fn parse_all_failing<P, T>(parser: P, programs: Vec<&str>) -> Vec<CustomDiagnostic>
+    pub(crate) fn parse_all_failing<P, T>(parser: P, programs: Vec<&str>) -> Vec<CustomDiagnostic>
     where
         P: NoirParser<T>,
         T: std::fmt::Display,
@@ -1748,13 +1701,13 @@ mod test {
     }
 
     #[derive(Copy, Clone)]
-    struct Case {
-        source: &'static str,
-        errors: usize,
-        expect: &'static str,
+    pub(crate) struct Case {
+        pub(crate) source: &'static str,
+        pub(crate) errors: usize,
+        pub(crate) expect: &'static str,
     }
 
-    fn check_cases_with_errors<T, P>(cases: &[Case], parser: P)
+    pub(crate) fn check_cases_with_errors<T, P>(cases: &[Case], parser: P)
     where
         P: NoirParser<T> + Clone,
         T: std::fmt::Display,
@@ -1791,6 +1744,13 @@ mod test {
 
         assert_eq!(vecmap(&results, |t| t.0.clone()), vecmap(&results, |t| t.1.clone()),);
     }
+}
+
+#[cfg(test)]
+mod test {
+    use super::parser_test_helpers::*;
+    use super::*;
+    use crate::{ArrayLiteral, Literal};
 
     #[test]
     fn regression_skip_comment() {
@@ -1962,142 +1922,6 @@ mod test {
     fn unwrap_expr(stmt: &StatementKind) -> &ExpressionKind {
         match stmt {
             StatementKind::Expression(expr) => &expr.kind,
-            _ => unreachable!(),
-        }
-    }
-
-    /// Deprecated constrain usage test
-    #[test]
-    fn parse_constrain() {
-        let errors = parse_with(constrain(expression()), "constrain x == y").unwrap_err();
-        assert_eq!(errors.len(), 1);
-        assert!(format!("{}", errors.first().unwrap()).contains("deprecated"));
-
-        // Currently we disallow constrain statements where the outer infix operator
-        // produces a value. This would require an implicit `==` which
-        // may not be intuitive to the user.
-        //
-        // If this is deemed useful, one would either apply a transformation
-        // or interpret it with an `==` in the evaluator
-        let disallowed_operators = vec![
-            BinaryOpKind::And,
-            BinaryOpKind::Subtract,
-            BinaryOpKind::Divide,
-            BinaryOpKind::Multiply,
-            BinaryOpKind::Or,
-        ];
-
-        for operator in disallowed_operators {
-            let src = format!("constrain x {} y;", operator.as_string());
-            let errors = parse_with(constrain(expression()), &src).unwrap_err();
-            assert_eq!(errors.len(), 2);
-            assert!(format!("{}", errors.first().unwrap()).contains("deprecated"));
-        }
-
-        // These are general cases which should always work.
-        //
-        // The first case is the most noteworthy. It contains two `==`
-        // The first (inner) `==` is a predicate which returns 0/1
-        // The outer layer is an infix `==` which is
-        // associated with the Constrain statement
-        let errors = parse_all_failing(
-            constrain(expression()),
-            vec![
-                "constrain ((x + y) == k) + z == y",
-                "constrain (x + !y) == y",
-                "constrain (x ^ y) == y",
-                "constrain (x ^ y) == (y + m)",
-                "constrain x + x ^ x == y | m",
-            ],
-        );
-        assert_eq!(errors.len(), 5);
-        assert!(errors
-            .iter()
-            .all(|err| { err.is_error() && err.to_string().contains("deprecated") }));
-    }
-
-    /// This is the standard way to declare an assert statement
-    #[test]
-    fn parse_assert() {
-        parse_with(assertion(expression()), "assert(x == y)").unwrap();
-
-        // Currently we disallow constrain statements where the outer infix operator
-        // produces a value. This would require an implicit `==` which
-        // may not be intuitive to the user.
-        //
-        // If this is deemed useful, one would either apply a transformation
-        // or interpret it with an `==` in the evaluator
-        let disallowed_operators = vec![
-            BinaryOpKind::And,
-            BinaryOpKind::Subtract,
-            BinaryOpKind::Divide,
-            BinaryOpKind::Multiply,
-            BinaryOpKind::Or,
-        ];
-
-        for operator in disallowed_operators {
-            let src = format!("assert(x {} y);", operator.as_string());
-            parse_with(assertion(expression()), &src).unwrap_err();
-        }
-
-        // These are general cases which should always work.
-        //
-        // The first case is the most noteworthy. It contains two `==`
-        // The first (inner) `==` is a predicate which returns 0/1
-        // The outer layer is an infix `==` which is
-        // associated with the Constrain statement
-        parse_all(
-            assertion(expression()),
-            vec![
-                "assert(((x + y) == k) + z == y)",
-                "assert((x + !y) == y)",
-                "assert((x ^ y) == y)",
-                "assert((x ^ y) == (y + m))",
-                "assert(x + x ^ x == y | m)",
-            ],
-        );
-
-        match parse_with(assertion(expression()), "assert(x == y, \"assertion message\")").unwrap()
-        {
-            StatementKind::Constrain(ConstrainStatement(_, message, _)) => {
-                let message = message.unwrap();
-                match message.kind {
-                    ExpressionKind::Literal(Literal::Str(message_string)) => {
-                        assert_eq!(message_string, "assertion message".to_owned());
-                    }
-                    _ => unreachable!(),
-                }
-            }
-            _ => unreachable!(),
-        }
-    }
-
-    /// This is the standard way to assert that two expressions are equivalent
-    #[test]
-    fn parse_assert_eq() {
-        parse_all(
-            assertion_eq(expression()),
-            vec![
-                "assert_eq(x, y)",
-                "assert_eq(((x + y) == k) + z, y)",
-                "assert_eq(x + !y, y)",
-                "assert_eq(x ^ y, y)",
-                "assert_eq(x ^ y, y + m)",
-                "assert_eq(x + x ^ x, y | m)",
-            ],
-        );
-        match parse_with(assertion_eq(expression()), "assert_eq(x, y, \"assertion message\")")
-            .unwrap()
-        {
-            StatementKind::Constrain(ConstrainStatement(_, message, _)) => {
-                let message = message.unwrap();
-                match message.kind {
-                    ExpressionKind::Literal(Literal::Str(message_string)) => {
-                        assert_eq!(message_string, "assertion message".to_owned());
-                    }
-                    _ => unreachable!(),
-                }
-            }
             _ => unreachable!(),
         }
     }
